AC: move YOLO adapters to separated file
